Improving sanity/consistency for integer types
All fixed width should be from stdint.h, with format strings from
inttypes.h, custom types with #define'd format strings, or size_t. Also
adding a function that prints an xint_t as UTF-8

+++ src/space.h
@@ -19,7 +19,7 @@ typedef struct space_t {
 } space_t;
 
 /* Create and initialize a new space */
-space_t *space_create(vector3_t block_size, unsigned long long hash_size);
+space_t *space_create(vector3_t block_size, size_t hash_size);
 /* Deallocate and destroy an existing space */
 void space_destroy(space_t *space);
 /* Set a cell in the space */
